+++ react/components/UserWidget.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import type { FunctionComponent } from 'react'
 import { useQuery } from 'react-apollo'
 import { useIntl, defineMessages } from 'react-intl'
-import { Button } from 'vtex.styleguide'
+import { Button, Tag } from 'vtex.styleguide'
 import { useCssHandles } from 'vtex.css-handles'
 import { useRuntime } from 'vtex.render-runtime'
 
@@ -23,6 +23,18 @@ const messages = defineMessages({
   },
   costCenter: {
     id: `${storePrefix}user-widget.costCenter`,
+  },
+  status: {
+    id: `${storePrefix}user-widget.status`,
+  },
+  active: {
+    id: `${storePrefix}user-widget.status.active`,
+  },
+  onHold: {
+    id: `${storePrefix}user-widget.status.on-hold`,
+  },
+  inactive: {
+    id: `${storePrefix}user-widget.status.inactive`,
   },
   manageOrganization: {
     id: `${storePrefix}user-widget.manage-organization`,
@@ -72,6 +84,34 @@ const UserWidget: FunctionComponent = () => {
     skip: !isAuthenticated,
   })
 
+  const handleStatusMessage = (status: string) => {
+    switch (status) {
+      case 'active':
+        return (
+          <Tag type="success" size="small">
+            {formatMessage(messages.active)}
+          </Tag>
+        )
+
+      case 'on-hold':
+        return (
+          <Tag type="warning" size="small">
+            {formatMessage(messages.onHold)}
+          </Tag>
+        )
+
+      case 'inactive':
+        return (
+          <Tag type="error" size="small">
+            {formatMessage(messages.inactive)}
+          </Tag>
+        )
+
+      default:
+        return ''
+    }
+  }
+
   if (
     !isAuthenticated ||
     !permissionsData ||
@@ -89,7 +129,10 @@ const UserWidget: FunctionComponent = () => {
       >
         {`${formatMessage(messages.organization)} ${
           organizationData?.getOrganizationByIdStorefront?.name
-        }`}
+        }`}{' '}
+        {handleStatusMessage(
+          organizationData?.getOrganizationByIdStorefront?.status ?? ''
+        )}
       </div>
       <div
         className={`${handles.userWidgetItem} pa3 br2 bg-base--inverted hover-bg-base--inverted active-bg-base--inverted c-on-base--inverted hover-c-on-base--inverted active-c-on-base--inverted dib mr3`}