+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Organization status now shown in `b2b-user-widget`
+
 ## [0.5.0] - 2021-12-10
 
 ### Added